+++ packages/react-components-storybook/src/stories/ObjectTable/Features/Selection.stories.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { ObjectTable } from "@osdk/react-components/experimental/object-table";
+import { ObjectTable } from "@osdk/react-components/object-table";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { useCallback, useState } from "react";
 import { expect, fn, userEvent, waitFor, within } from "storybook/test";