+++ packages/react-components/src/public/experimental/action-form.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Palantir Technologies, Inc. All rights reserved.
+ * Copyright 2026 Palantir Technologies, Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,40 +14,106 @@
  * limitations under the License.
  */
 
-import { ActionForm as _ActionForm } from "../../action-form/ActionForm.js";
-import { withOsdkMetrics } from "../../util/withOsdkMetrics.js";
-export const ActionForm: typeof _ActionForm = withOsdkMetrics(
-  _ActionForm,
-  "ActionForm",
-);
-export type {
-  ActionFormProps,
-  BaseFormProps,
-  FormContentItem,
-  FormError,
-  FormSectionDefinition,
-  FormState,
-} from "../../action-form/ActionFormApi.js";
-export { BaseForm } from "../../action-form/BaseForm.js";
-export type {
-  ActionParameters,
-  BaseFormFieldProps,
-  CustomFieldProps,
-  DropdownFieldProps,
-  FieldComponent,
-  FieldValueType,
-  FilePickerProps,
-  FormFieldDefinition,
-  FormFieldPropsByType,
-  NumberInputFieldProps,
-  ObjectSelectFieldProps,
-  ObjectSetFieldProps,
-  Option,
-  PortalContainer,
-  RadioButtonsFieldProps,
-  RendererFieldDefinition,
-  TextAreaFieldProps,
-  TextInputFieldProps,
-  UnsupportedFieldProps,
-  ValidationError,
-} from "../../action-form/FormFieldApi.js";
+import type * as Api from "@osdk/api";
+
+import type * as FormFieldApi from "../../action-form/FormFieldApi.js";
+import * as Stable from "../action-form.js";
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export const ActionForm: typeof Stable.ActionForm = Stable.ActionForm;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type ActionFormProps<Q extends Api.ActionDefinition<unknown>> =
+  Stable.ActionFormProps<Q>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type BaseFormProps = Stable.BaseFormProps;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FormContentItem = Stable.FormContentItem;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FormError = Stable.FormError;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FormSectionDefinition = Stable.FormSectionDefinition;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FormState<Q extends Api.ActionDefinition<unknown>> =
+  Stable.FormState<Q>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export const BaseForm: typeof Stable.BaseForm = Stable.BaseForm;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type ActionParameters<Q extends Api.ActionDefinition<unknown>> =
+  Stable.ActionParameters<Q>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type BaseFormFieldProps<V> = Stable.BaseFormFieldProps<V>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type CustomFieldProps<V> = Stable.CustomFieldProps<V>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type DropdownFieldProps<
+  V,
+  Multiple extends boolean = false,
+> = Stable.DropdownFieldProps<V, Multiple>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FieldComponent = Stable.FieldComponent;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FieldValueType<
+  Q extends Api.ActionDefinition<unknown>,
+  K extends keyof Stable.ActionParameters<Q> = keyof Stable.ActionParameters<Q>,
+> = Stable.FieldValueType<Q, K>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FilePickerProps = Stable.FilePickerProps;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FormFieldDefinition<
+  Q extends Api.ActionDefinition<unknown>,
+  K extends FormFieldApi.FieldKey<Q> = FormFieldApi.FieldKey<Q>,
+> = Stable.FormFieldDefinition<Q, K>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type FormFieldPropsByType = Stable.FormFieldPropsByType;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type NumberInputFieldProps = Stable.NumberInputFieldProps;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type ObjectSelectFieldProps<
+  Q extends Api.ObjectTypeDefinition = Api.ObjectTypeDefinition,
+> = Stable.ObjectSelectFieldProps<Q>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type ObjectSetFieldProps<T extends Api.ObjectTypeDefinition> =
+  Stable.ObjectSetFieldProps<T>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type Option<V> = Stable.Option<V>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type PortalContainer = Stable.PortalContainer;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type RadioButtonsFieldProps<V> = Stable.RadioButtonsFieldProps<V>;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type RendererFieldDefinition = Stable.RendererFieldDefinition;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type TextAreaFieldProps = Stable.TextAreaFieldProps;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type TextInputFieldProps = Stable.TextInputFieldProps;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type UnsupportedFieldProps = Stable.UnsupportedFieldProps;
+
+/** @deprecated Import from `@osdk/react-components/action-form` instead. */
+export type ValidationError = Stable.ValidationError;