+++ packages/create-app.template.tutorial-todo-app.beta/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.tutorial-todo-app.beta
 
+## 2.66.0
+
+## 2.65.0
+
 ## 2.64.0
 
 ### Minor Changes