+++ packages/create-app.template.typescript-library.beta/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.typescript-library.beta
 
+## 2.66.0
+
+## 2.65.0
+
 ## 2.64.0
 
 ## 2.63.0