+++ packages/create-app.template.vue.v2/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.vue.v2
 
+## 2.66.0
+
+## 2.65.0
+
 ## 2.64.0
 
 ### Minor Changes