+++ packages/create-widget.template.react.v2/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-widget.template.react.v2
 
+## 3.65.0
+
+## 3.64.0
+
 ## 3.63.0
 
 ## 3.62.0