+++ packages/create-widget/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/create-widget
 
+## 3.65.0
+
+### Patch Changes
+
+- @osdk/generator-utils@2.66.0
+
+## 3.64.0
+
+### Patch Changes
+
+- @osdk/generator-utils@2.65.0
+
 ## 3.63.0
 
 ### Patch Changes