+++ packages/react-components-storybook/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @osdk/react-components-storybook
 
+## 0.54.0
+
+### Patch Changes
+
+- @osdk/faux@0.50.0
+
+## 0.53.0
+
+### Patch Changes
+
+- Updated dependencies [e3a381d]
+  - @osdk/faux@0.49.0
+
 ## 0.52.0
 
 ### Minor Changes