+++ packages/widget.api/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @osdk/widget.api
 
+## 3.65.0
+
+### Patch Changes
+
+- @osdk/api@2.66.0
+
+## 3.64.0
+
+### Patch Changes
+
+- Updated dependencies [e3a381d]
+  - @osdk/api@2.65.0
+
 ## 3.63.0
 
 ### Patch Changes